fix arcosh clamping
In the `arccosh` function of the math.py file for the Lorentz manifold, the clamping for input vector $x$ was not correct.

The original code tried to enforce the domain of arcosh with `z = torch.clamp_min(x.double().pow(2) - 1.0, 1e-15))`. However, in the next line of code, we compute `torch.log(x+z)`. Thus although the variable `z` was valid, the variable `x` still was not. This meant that any $x<0$ would cause a `nan` error instead of being clamped appropriately.

This commit instead simply redefines `x = torch.clamp_min(x, 1.0).double()` thus resolving the issue.